Rename __tests__ dir to tests
